requirements.txt
discord
discord.py
pillow
requests